refactor csods api interface types directory
