+++ client/src/features/project-details/ProjectDetails.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { DEFAULT_USER } from "../../constants/defaults";
-import { IAllProjectTags, IProjectDetails, IUser } from "../../viewModels/csods/csodsApiInterfaces";
+import { IUser, IProjectDetails, IAllProjectTags } from "@/types";
 import ApiHandler from "../../utils/api/ApiHandler";
 import CoreContributors from "./components/CoreContributors";
 import ProjectDetailsProvider from "./components/ProjectDetailsProvider";
