+++ client/src/features/project-details/components/ProjectInformation.tsx
@@ -2,7 +2,7 @@ import { createContext, useContext, useEffect, useMemo, useState } from "react";
 import { Color, OnHover, Opacity } from "../../../types";
 import { ICONS } from "../../../constants/project_data/tags";
 import { PROJECT_DESCRIPTION } from "../../../constants/defaults";
-import { IProjectTags } from "../../../viewModels/csods/csodsApiInterfaces";
+import { IProjectTags } from "../../../types";
 import { AllTagsContext, ProjectContext, UserContext } from "./ProjectDetailsProvider";
 import { ProjectDataService } from "../../../utils/data/ProjectDataService";
 import TagRow from "../../../components/shared/TagRow";
