+++ client/src/types/api/csods/projects/projectsPage.ts
@@ -1,0 +1,9 @@
+import { IProjectDetails } from "./project"
+
+export interface IProjectsPage {
+    TotalPages: number,
+    Projects: IProjectDetails[]
+};
+
+
+
