+++ client/src/types/api/csods/user/user.ts
@@ -1,0 +1,6 @@
+export interface IUser {
+    Name: string,
+    Email: string
+};
+
+
