+++ client/src/types/index.ts
@@ -1,1 +1,5 @@
+export type { IUser } from './api/csods/user/user';
+export type { IProjectDetails, IProject, IProjectTags } from './api/csods/projects/project';
+export type { IAllProjectTags, IDevType, IProgrammingLanguage, IFramework, IProjectFramework, IDatabaseTechnology, IApplicationIndustry } from './api/csods/projects/tags/projectTags';
+export type { IProjectsPage } from './api/csods/projects/projectsPage';
 export type { Opacity, Color, OnHover } from './themes/styles';
